requirements.txt
streamlit
openai
pypdf
python-docx